+++ src/Routes/Pages/Category/Components/index.js
@@ -1,0 +1,1 @@
+export { default as CategoryModal } from './CategoryModal';
